requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
python-dotenv==1.0.0
langchain==0.0.200
langgraph==0.0.10
openai==0.28.0
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
SQLAlchemy==2.0.12
alembic==1.11.2
psycopg2-binary[stderr]==2.9.6
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==1.3.1
